Allow the dev environment to start, with warnings, if the "secrets" repository is not found
I don’t have access to Knightlab’s secrets - so a lot of the dev
environment breaks. It appears that what comes up on localhost is
served via the Knightlab S3 bucket rather than localhost anyway; this
is confusing; but these changes at least allow the server to run so
that we can point to it to serve out rebuilt JS files. I guess?